Lumberyard Release 1.21.0.0
        Binaries Uncompressed Size: 21.69GB (23292944251 bytes)

+++ dev/Code/Framework/AzQtComponents/AzQtComponents/Components/TitleBarOverdrawHandler.cpp
@@ -242,7 +242,7 @@ void TitleBarOverdrawHandlerWindows::applyOverdrawMargins(QPlatformWindow* windo
         {
             const auto exStyle = GetWindowLongPtr(hWnd, GWL_EXSTYLE);
             const auto monitor = MonitorFromWindow(hWnd, MONITOR_DEFAULTTONULL);
-            const auto margins = customTitlebarMargins(monitor, style, exStyle, maximized);
+            const auto margins = customTitlebarMargins(monitor, static_cast<int>(style), static_cast<int>(exStyle), maximized);
             RECT rect;
             GetWindowRect(hWnd, &rect);
             pni->setWindowProperty(window, QStringLiteral("WindowsCustomMargins"), qVariantFromValue(margins));
